insurance_streamlit/requirements.txt
streamlit==1.48.1
snowflake-ml-python==1.10.0
snowflake-snowpark-python==1.37.0